requirements.txt
streamlit>=1.32.0
gspread>=6.0.0
google-auth>=2.28.0
pandas>=2.2.0
plotly>=5.20.0
numpy>=1.26.0
